feat: Add language check for browser language in App component
This commit adds a language check for the browser language in the App component. Previously, the English language was set as the default if the browser language was not \hu-HU\. However, this caused an issue for users whose browser language was set to \hu\. This commit updates the language check to exclude both \hu-HU\ and \hu\ from setting English as the default language.